--- .env.example
@@ -1,16 +1,0 @@
-# Hugging Face API Token (Required)
-# Get from: https://huggingface.co/settings/tokens
-# Choose "Read" permissions
-HF_TOKEN=your_hugging_face_token_here
-
-# GitHub Personal Access Token (Optional)
-# Get from: https://github.com/settings/tokens
-# Select scopes: public_repo, read:user
-# This increases API rate limits and provides access to private repos
-GITHUB_TOKEN=your_github_token_here
-
-# Google Gemini API Key (Required for LLM functionality)
-# Get from: https://console.cloud.google.com/apis/credentials
-# Enable the Generative AI API first, then create an API key
-# Alternative: https://aistudio.google.com/app/apikey (Google AI Studio)
-GEMINI_API_KEY=your_gemini_api_key_here
