--- src/Commands/Cli.py
@@ -1,212 +1,0 @@
-import argparse
-import sys
-from typing import List, Optional
-
-
-def handle_install(args) -> int:
-    """
-    Handle the install subcommand.
-    Sets up necessary dependencies and configurations.
-    
-    Args:
-        args: Parsed arguments from argparse
-        
-    Returns:
-        int: Exit code (0 for success, non-zero for failure)
-    """
-    print("Installing dependencies and setting up the environment...")
-    
-    # TODO: Implement actual installation logic
-    # This might include:
-    # - Installing required Python packages
-    # - Setting up API keys/tokens
-    # - Creating configuration files
-    # - Validating environment setup
-    
-    print("Installation completed successfully!")
-    return 0
-
-
-def handle_run(args) -> int:
-    """
-    Handle the run subcommand.
-    Main functionality to analyze models and generate scores.
-    
-    Args:
-        args: Parsed arguments from argparse
-        
-    Returns:
-        int: Exit code (0 for success, non-zero for failure)
-    """
-    print(f"Running analysis on URL file: {args.url_file}")
-    
-    if args.net_score:
-        print("Net score calculation enabled")
-    
-    # TODO: Implement actual run logic
-    # This will include:
-    # - Reading URLs from the input file
-    # - Classifying URLs (model/dataset/code)
-    # - Fetching metadata from APIs
-    # - Calculating metrics in parallel
-    # - Computing net scores
-    # - Outputting results in NDJSON format
-    
-    print("Analysis completed!")
-    return 0
-
-
-def handle_test(args) -> int:
-    """
-    Handle the test subcommand.
-    Runs the test suite to validate functionality.
-    
-    Args:
-        args: Parsed arguments from argparse
-        
-    Returns:
-        int: Exit code (0 for success, non-zero for failure)
-    """
-    print("Running test suite...")
-    
-    if args.coverage:
-        print("Running tests with coverage analysis")
-    
-    # TODO: Implement test execution logic
-    # This will include:
-    # - Running pytest with coverage
-    # - Validating >= 80% code coverage
-    # - Running >= 20 distinct tests
-    # - Reporting test results
-    
-    print("All tests passed!")
-    return 0
-
-
-def create_parser() -> argparse.ArgumentParser:
-    """
-    Create and configure the argument parser with subcommands.
-    
-    Returns:
-        argparse.ArgumentParser: Configured parser
-    """
-    # Main parser
-    parser = argparse.ArgumentParser(
-        prog="trustworthy-model-cli",
-        description="CLI tool for cataloging and scoring pre-trained models",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog="""
-Examples:
-  %(prog)s install
-  %(prog)s run sample_urls.txt
-  %(prog)s run sample_urls.txt --net-score
-  %(prog)s test
-  %(prog)s test --coverage
-        """
-    )
-    
-    # Add version argument
-    parser.add_argument(
-        "--version", 
-        action="version", 
-        version="%(prog)s 1.0.0"
-    )
-    
-    # Create subparsers
-    subparsers = parser.add_subparsers(
-        dest="command",
-        help="Available commands",
-        metavar="COMMAND"
-    )
-    
-    # Install subcommand
-    install_parser = subparsers.add_parser(
-        "install",
-        help="Install dependencies and set up the environment",
-        description="Set up the environment and install necessary dependencies"
-    )
-    install_parser.set_defaults(func=handle_install)
-    
-    # Run subcommand
-    run_parser = subparsers.add_parser(
-        "run",
-        help="Analyze models from URL file and generate scores",
-        description="Run analysis on models specified in the URL file"
-    )
-    run_parser.add_argument(
-        "url_file",
-        help="Path to file containing URLs to analyze (one per line)"
-    )
-    run_parser.add_argument(
-        "--net-score",
-        action="store_true",
-        help="Calculate and include net scores in output"
-    )
-    run_parser.add_argument(
-        "--output",
-        "-o",
-        help="Output file path (default: stdout)",
-        default=None
-    )
-    run_parser.add_argument(
-        "--parallel",
-        "-p",
-        type=int,
-        help="Number of parallel workers (default: 4)",
-        default=4
-    )
-    run_parser.set_defaults(func=handle_run)
-    
-    # Test subcommand
-    test_parser = subparsers.add_parser(
-        "test",
-        help="Run the test suite",
-        description="Execute tests to validate functionality"
-    )
-    test_parser.add_argument(
-        "--coverage",
-        action="store_true",
-        help="Run tests with coverage analysis"
-    )
-    test_parser.add_argument(
-        "--verbose",
-        "-v",
-        action="store_true",
-        help="Verbose test output"
-    )
-    test_parser.set_defaults(func=handle_test)
-    
-    return parser
-
-
-def main(argv: Optional[List[str]] = None) -> int:
-    """
-    Main entry point for the CLI application.
-    
-    Args:
-        argv: Command line arguments (uses sys.argv if None)
-        
-    Returns:
-        int: Exit code
-    """
-    parser = create_parser()
-    args = parser.parse_args(argv)
-    
-    # If no subcommand provided, show help
-    if not hasattr(args, 'func'):
-        parser.print_help()
-        return 1
-    
-    try:
-        # Call the appropriate handler function
-        return args.func(args)
-    except KeyboardInterrupt:
-        print("\nOperation cancelled by user", file=sys.stderr)
-        return 130
-    except Exception as e:
-        print(f"Error: {e}", file=sys.stderr)
-        return 1
-
-
-if __name__ == "__main__":
-    sys.exit(main())
