--- src/Commands/__init__.py
@@ -1,1 +1,0 @@
-"""Commands module for CLI functionality."""
