--- src/Controllers/__init__.py
@@ -1,1 +1,0 @@
-"""Controllers module for MVC pattern."""
