--- src/Models/__init__.py
@@ -1,1 +1,0 @@
-"""Models module for data structures."""
