--- src/Testing/__init__.py
@@ -1,1 +1,0 @@
-"""Testing module for unit tests."""
