--- src/__init__.py
@@ -1,3 +1,0 @@
-"""
-Trustworthy Model CLI Package
-"""
